requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14,<3.15
django-cors-headers>=4.3,<4.4